fix ordinals content parsing
