+++ api/index.ts
@@ -2,3 +2,4 @@ export * from './btc';
 export * from './xverse';
 export * from './stacks';
 export * from './gamma';
+export * from './ordinals';
